system/requirements.txt
pyserial
pyvisa
pyvisa-py
psutil
zeroconf
numpy
flask
flask_cors
flask_restful
